Redesign HistoryDict internal storage, fix tests
HistoryDict internal storage was redisigned with a simpler homogeneous structure.

It was: {<key>: {
	'newest': (<timestamp>, <item>),
	<timestamp>: <item>,
	...
}}

It became: {<key>: {
	<timestamp>: <item>,
	<timestamp>: <item>,
	...
}}

Newest field was removed and the order of the elements of history was taken into account.
From now on older elements are placed at the beginning, and newer - at the end of internal dict